+++ App/io_bus.cpp
@@ -22,11 +22,10 @@
  * @brief Initialize I2C bus by default
  *
  */
-IOError IOBus::Initialize(void *ctx)
-{
+IOError IOBus::Initialize(void *ctx) {
     UNUSED(ctx);
     LOG_INFO("I2C bus init");
-    return  IOError::kIO_OK;
+    return IOError::kIO_OK;
 }
 
 /**
@@ -35,8 +34,7 @@ IOError IOBus::Initialize(void *ctx)
  * @param adr
  * @return I2CBusError
  */
-IOError IOBus::IsReady(uint8_t adr)
-{
+IOError IOBus::IsReady(uint8_t adr) {
     if (adr != kVirtualI2CDeviceAddress)
         return IOError::kIO_ERR;
     LOG_INFO("Find I2C virtual: %#x", adr);
@@ -54,8 +52,7 @@ IOError IOBus::IsReady(uint8_t adr)
  * @return I2CBusError
  */
 IOError IOBus::Write(uint8_t adr, uint8_t reg, uint8_t regsize,
-                            uint8_t *data, uint8_t len)
-{
+                     uint8_t *data, uint8_t len) {
     LOG_INFO("Write I2C bus: ADR: %#x REG: %#x REGSIZE: %d", adr, reg, regsize);
     LOG_RAW_INFO("DATA[%d]: ", len);
     for (int32_t i = 0; i < len; ++i)
@@ -76,12 +73,10 @@ IOError IOBus::Write(uint8_t adr, uint8_t reg, uint8_t regsize,
  * @return I2CBusError
  */
 IOError IOBus::Read(uint8_t adr, uint8_t reg, uint8_t regsize,
-                           uint8_t *data, uint8_t len)
-{
+                    uint8_t *data, uint8_t len) {
     LOG_INFO("Read I2C bus: ADR: %#x REG: %#x REGSIZE: %d", adr, reg, regsize);
     LOG_RAW_INFO("DATA[%d]: ", len);
-    for (int32_t i = 0; i < len; ++i)
-    {
+    for (int32_t i = 0; i < len; ++i) {
         data[i] = i;
         LOG_RAW_INFO("%#x ", data[i]);
     }
@@ -95,8 +90,7 @@ IOError IOBus::Read(uint8_t adr, uint8_t reg, uint8_t regsize,
  * @param err
  * @return const char*
  */
-const char *IOBus::ErrStringify(IOError err)
-{
+const char *IOBus::ErrStringify(IOError err) {
     if (err == kIO_TIMEOUT)
         return "IO_I2C_TIMEOUT";
     else if (err == kIO_BUSY)