+++ App/platform/core407/core407bus.cpp
@@ -11,6 +11,7 @@
  *
  */
 #include "core407bus.hpp"
+
 #include "i2c.h"
 //>>---------------------- Log control
 #define LOG_MODULE_NAME c407bus
@@ -21,8 +22,7 @@
 #endif
 #include "log_libs.h"
 //<<----------------------
-IOError Core407Bus::to_iobus_error(int err)
-{
+IOError Core407Bus::to_iobus_error(int err) {
     if (err == HAL_ERROR)
         return kIO_ERR;
     else if (err == HAL_BUSY)
@@ -38,8 +38,7 @@ IOError Core407Bus::to_iobus_error(int err)
  * @param ctx
  * @return IOBusError
  */
-IOError Core407Bus::Initialize(void *ctx)
-{
+IOError Core407Bus::Initialize(void *ctx) {
     i2c_handle_ = &hi2c1;
     return kIO_OK;
 }
@@ -50,8 +49,7 @@ IOError Core407Bus::Initialize(void *ctx)
  * @param adr
  * @return IOBusError
  */
-IOError Core407Bus::IsReady(uint8_t adr)
-{
+IOError Core407Bus::IsReady(uint8_t adr) {
     HAL_StatusTypeDef err =
         HAL_I2C_IsDeviceReady(i2c_handle_, (adr << 1), 1, kI2CTimeoutMs);
     return Core407Bus::to_iobus_error(err);
@@ -68,8 +66,7 @@ IOError Core407Bus::IsReady(uint8_t adr)
  * @return IOBusError
  */
 IOError Core407Bus::Write(uint8_t adr, uint8_t reg, uint8_t regsize,
-                             uint8_t *data, uint8_t len)
-{
+                          uint8_t *data, uint8_t len) {
     HAL_StatusTypeDef err = HAL_I2C_Mem_Write(
         i2c_handle_, (adr << 1), reg, regsize, data, len, kI2CTimeoutMs);
     return Core407Bus::to_iobus_error(err);
@@ -86,8 +83,7 @@ IOError Core407Bus::Write(uint8_t adr, uint8_t reg, uint8_t regsize,
  * @return IOBusError
  */
 IOError Core407Bus::Read(uint8_t adr, uint8_t reg, uint8_t regsize,
-                            uint8_t *data, uint8_t len)
-{
+                         uint8_t *data, uint8_t len) {
     HAL_StatusTypeDef err = HAL_I2C_Mem_Read(i2c_handle_, (adr << 1), reg,
                                              regsize, data, len, kI2CTimeoutMs);
     return Core407Bus::to_iobus_error(err);