+++ App/platform/core407/core407io.hpp
@@ -2,14 +2,13 @@
 #ifndef CORE407IO_H
 #define CORE407IO_H
 
-#include "sys_gpio.hpp"
 #include "errors.h"
+#include "sys_gpio.hpp"
 
-class Core407IO : public SysGpio
-{
-  public:
+class Core407IO : public SysGpio {
+   public:
     IOError Initialize(void *pv);
     IOError LedWrite(bool state);
 };
 
-#endif // CORE407IO_H
+#endif  // CORE407IO_H