+++ App/platform/green_pill/gp_io.hpp
@@ -2,14 +2,13 @@
 #ifndef GP_IO_H
 #define GP_IO_H
 
-#include "sys_gpio.hpp"
 #include "errors.h"
+#include "sys_gpio.hpp"
 
-class GpIO : public SysGpio
-{
-  public:
+class GpIO : public SysGpio {
+   public:
     IOError Initialize(void *pv);
     IOError LedWrite(bool state);
 };
 
-#endif // GP_IO_H
+#endif  // GP_IO_H