+++ App/platform/green_pill/green_pill.hpp
@@ -6,16 +6,15 @@
 #include "gp_io.hpp"
 #include "system.hpp"
 
-class GreenPill : public System
-{
-  public:
+class GreenPill : public System {
+   public:
     SystemError Initialize(void *ctx);
     IOBus *GetI2CBus();
     SysGpio *GetIo();
 
-  private:
+   private:
     GpBus i2c_bus_;
     GpIO io_;
 };
 
-#endif // CORE407_H
+#endif  // CORE407_H