+++ CubeMX/STM32F401CxUx/USB_DEVICE/App/usbd_cdc_if.c
@@ -23,7 +23,7 @@
 #include "usbd_cdc_if.h"
 
 /* USER CODE BEGIN INCLUDE */
-#include "cli_handle.h"
+#include "ringbuffer.h"
 /* USER CODE END INCLUDE */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -110,7 +110,7 @@ uint8_t UserTxBufferFS[APP_TX_DATA_SIZE];
 extern USBD_HandleTypeDef hUsbDeviceFS;
 
 /* USER CODE BEGIN EXPORTED_VARIABLES */
-
+extern ring_buffer_t input_ring;
 /* USER CODE END EXPORTED_VARIABLES */
 
 /**
@@ -265,9 +265,7 @@ static int8_t CDC_Receive_FS(uint8_t* Buf, uint32_t *Len)
 {
   /* USER CODE BEGIN 6 */
   USBD_CDC_SetRxBuffer(&hUsbDeviceFS, &Buf[0]);
-  for (uint32_t cnt = 0; cnt < *Len; cnt++){
-    CliPutToBuf(Buf[cnt]);
-  }
+  ring_buffer_queue_arr(&input_ring, (const char*)Buf, *Len);
   USBD_CDC_ReceivePacket(&hUsbDeviceFS);
   return (USBD_OK);
   /* USER CODE END 6 */